+++ src/components/SideNavBar/index.jsx
@@ -2,7 +2,7 @@ import SideNavItem from "./SideNavItem";
 import "./SideNavBar.style.css";
 import { memo } from "react";
 import { useSelector } from "react-redux";
-import { selectSideNavCategoryIds } from "../../state/reducers/sideNavCategoriesReducer";
+import { selectSideNavCategoryIds } from "../../state/reducers/selectors/sideNavCategories";
 
 const SideNavBar = () => {
   const categoryIds = useSelector(selectSideNavCategoryIds);
@@ -11,12 +11,7 @@ const SideNavBar = () => {
     <div className="side__nav">
       {categoryIds.length > 0 &&
         categoryIds.map((categoryId) => {
-          return (
-            <SideNavItem
-              key={categoryId}
-              categoryId={categoryId}
-            />
-          );
+          return <SideNavItem key={categoryId} categoryId={categoryId} />;
         })}
     </div>
   );
