+++ src/state/action-creators/index.js
@@ -1,56 +1,56 @@
-import ActionType from "../action-types";
+import ACTION_TYPE from "../action-types";
 
 /* Action Creators for Products */
 export const productsLoading = () => ({
-  type: ActionType.PRODUCTS_LOADING,
+  type: ACTION_TYPE.PRODUCTS_LOADING,
 });
 
 export const productsLoaded = (products) => ({
-  type: ActionType.PRODUCTS_LOADED,
+  type: ACTION_TYPE.PRODUCTS_LOADED,
   payload: products,
 });
 
 export const productsOrderChange = (order) => ({
-  type: ActionType.PRODUCTS_ORDER_CHANGE,
+  type: ACTION_TYPE.PRODUCTS_ORDER_CHANGE,
   payload: order,
 });
 
 /* Action Creators for Side Nav Categories */
 export const sideNavCategoriesLoading = () => ({
-  type: ActionType.SIDE_NAV_CATEGORIES_LOADING,
+  type: ACTION_TYPE.SIDE_NAV_CATEGORIES_LOADING,
 });
 
 export const sideNavCategoriesLoaded = (categories) => ({
-  type: ActionType.SIDE_NAV_CATEGORIES_LOADED,
+  type: ACTION_TYPE.SIDE_NAV_CATEGORIES_LOADED,
   payload: categories,
 });
 
-export const sideNavCurrentCategoryChanged = (categoryId) => ({
-  type: ActionType.SIDE_NAV_CURRENT_CATEGORY_CHANGED,
+export const changeSideNavCategory = (categoryId) => ({
+  type: ACTION_TYPE.SIDE_NAV_CURRENT_CATEGORY_CHANGED,
   payload: categoryId,
 });
 
 /* Action Creators for Cart */
 export const addProductToCart = (productId) => {
   return {
-    type: ActionType.PRODUCT_ADDED_TO_CART,
+    type: ACTION_TYPE.PRODUCT_ADDED_TO_CART,
     payload: productId,
   };
 };
 
 export const removeProductFromCart = (productId) => {
   return {
-    type: ActionType.PRODUCT_REMOVED_FROM_CART,
+    type: ACTION_TYPE.PRODUCT_REMOVED_FROM_CART,
     payload: productId,
   };
 };
 
 /* Action Creators for sub-header categories */
 export const subHeaderCategoriesLoading = () => ({
-  type: ActionType.SUB_HEADER_CATEGORIES_LOADING,
+  type: ACTION_TYPE.SUB_HEADER_CATEGORIES_LOADING,
 });
 
 export const subHeaderCategoriesLoaded = (categories) => ({
-  type: ActionType.SUB_HEADER_CATEGORIES_LOADED,
+  type: ACTION_TYPE.SUB_HEADER_CATEGORIES_LOADED,
   payload: categories,
 });
