+++ src/state/action-types/index.js
@@ -1,4 +1,4 @@
-const ActionType = {
+const ACTION_TYPE = {
   PRODUCTS_LOADED: "products/productsLoaded",
   PRODUCTS_LOADING: "products/productsLoading",
   PRODUCTS_ORDER_CHANGE: "products/productsOrderChange",
@@ -6,11 +6,10 @@ const ActionType = {
   PRODUCT_REMOVED_FROM_CART: "cart/productRemovedFromCart",
   SIDE_NAV_CATEGORIES_LOADED: "sideNavCategories/sideNavCategoriesLoaded",
   SIDE_NAV_CATEGORIES_LOADING: "sideNavCategories/sideNavCategoriesLoading",
-  SIDE_NAV_CURRENT_CATEGORY_CHANGED:
-    "sideNavCategories/sideNavCurrentCategoryChanged",
+  SIDE_NAV_CURRENT_CATEGORY_CHANGED: "sideNavCategories/changeSideNavCategory",
   SUB_HEADER_CATEGORIES_LOADED: "subHeaderCategories/subHeaderCategoriesLoaded",
   SUB_HEADER_CATEGORIES_LOADING:
     "subHeaderCategories/subHeaderCategoriesLoading",
 };
 
-export default ActionType;
+export default ACTION_TYPE;
