+++ src/state/reducers/productsReducer.js
@@ -1,6 +1,4 @@
-import { createSelector } from "reselect";
-import { SORTING_ORDER } from "../../assets/constants";
-import ActionType from "../action-types";
+import ACTION_TYPE from "../action-types";
 
 const initialState = {
   status: "idle",
@@ -10,7 +8,7 @@ const initialState = {
 
 export default function productsReducer(state = initialState, action) {
   switch (action.type) {
-    case ActionType.PRODUCTS_LOADED:
+    case ACTION_TYPE.PRODUCTS_LOADED:
       const newProducts = {};
 
       action.payload.forEach((product) => {
@@ -23,13 +21,13 @@ export default function productsReducer(state = initialState, action) {
         entities: newProducts,
       };
 
-    case ActionType.PRODUCTS_LOADING:
+    case ACTION_TYPE.PRODUCTS_LOADING:
       return {
         ...state,
         status: "loading",
       };
 
-    case ActionType.PRODUCTS_ORDER_CHANGE:
+    case ACTION_TYPE.PRODUCTS_ORDER_CHANGE:
       return {
         ...state,
         order: action.payload,
@@ -39,38 +37,3 @@ export default function productsReducer(state = initialState, action) {
       return state;
   }
 }
-
-const selectProductEntities = (state) => state.products.entities;
-
-export const selectProducts = createSelector(
-  // input selectors
-  selectProductEntities,
-  (state) => state.products.order,
-  // output selector
-  (entities, order) => {
-    const products = Object.values(entities);
-    if (order === SORTING_ORDER.LOWEST_PRICE) {
-      return products.sort((a, b) => a.discountedPrice - b.discountedPrice);
-    } else if (order === SORTING_ORDER.HIGHEST_PRICE) {
-      return products.sort((a, b) => b.discountedPrice - a.discountedPrice);
-    } else {
-      return products;
-    }
-  }
-);
-
-export const selectCurrentOrder = (state) => {
-  return state.products.order;
-};
-
-export const selectProductById = (state, productId) => {
-  return selectProductEntities(state)[productId];
-};
-
-// select all products of a categoryId
-export const selectProductsIdByCategoryId = (state, categoryId) => {
-  const products = selectProducts(state);
-  return products
-    .filter((product) => product.categoryId === categoryId)
-    .map((product) => product.productId);
-};
